Add tests for service classes and object types
