+++ scan-command/src/test/resources/command-outputs/unix/list-rules-output.txt
@@ -1,9 +1,10 @@
 Loading scan tool configurations from src/test/resources/test-resources/bal-project-with-config-file/Scan.toml
-	RuleID                                           | Rule Kind     | Rule Description                            
-	----------------------------------------------------------------------------------------------------------------
+	RuleID                                           | Rule Kind     | Rule Description                      
+	----------------------------------------------------------------------------------------------------------
 	ballerina:1                                      | CODE_SMELL    | Avoid checkpanic
 	ballerina:2                                      | CODE_SMELL    | Unused function parameter
-	ballerina:3                                      | CODE_SMELL    | Non isolated public class or function/method
+	ballerina:3                                      | CODE_SMELL    | Non isolated public function or method
+	ballerina:4                                      | CODE_SMELL    | Non isolated public class or object
 	ballerina/example_module_static_code_analyzer:1  | CODE_SMELL    | rule 1
 	ballerina/example_module_static_code_analyzer:2  | BUG           | rule 2
 	ballerina/example_module_static_code_analyzer:3  | VULNERABILITY | rule 3
